mod id type in base service

+++ domain-layer/src/main/java/com/wanderfun/domainlayer/repository/place/PlaceCategoryRepository.java
@@ -5,8 +5,8 @@ import com.wanderfun.domainlayer.repository.BaseRepository;
 
 import java.util.Optional;
 
-public interface PlaceCategoryRepository extends BaseRepository<PlaceCategory, Long> {
+public interface PlaceCategoryRepository extends BaseRepository<PlaceCategory, Integer> {
     Optional<PlaceCategory> findByName(String name);
     Optional<PlaceCategory> findByNameEn(String nameEn);
-    Optional<PlaceCategory> findById(Long id);
+    Optional<PlaceCategory> findById(Integer id);
 }
